Move WebPathField to own package
Pathfields can be used outside of filechoosers too,
so it makes sense to have a separate package for them.
