+++ modules/ui/src/com/alee/extended/pathfield/icons/down.png
