+++ modules/ui/src/com/alee/extended/pathfield/icons/left.png
