+++ modules/ui/src/com/alee/extended/pathfield/icons/right.png
